+++ database/migrations/2017_12_26_232437_create_pants_table.php
@@ -32,7 +32,7 @@ class CreatePantsTable extends Migration
 
             // Datos Pantalón
             $table->integer('fit_id')->nullable();
-            $table->integer('talla')->nullable();
+            $table->double('talla')->nullable();
 
             $table->string('notas',1000)->nullable();
